LPS-54322 Move EditPasswordPolicyAction.java, no logic changes
